requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.6.0
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
Babel==2.15.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
branca==0.8.1
cachetools==5.4.0
certifi==2024.2.2
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
chromedriver-autoinstaller==0.6.4
click==8.1.7
comm==0.2.2
contourpy==1.2.1
cryptography==43.0.3
cssselect==1.2.0
cssutils==2.11.1
cycler==0.12.1
Cython==3.0.11
dash==2.17.1
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses-json==0.6.7
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.18
distro==1.9.0
emoji==2.14.1
et-xmlfile==1.1.0
eval_type_backport==0.2.2
executing==2.0.1
faiss-cpu==1.9.0
fastjsonschema==2.20.0
filelock==3.15.4
filetype==1.2.0
Flask==3.0.3
folium==0.19.3
fonttools==4.52.1
formulaic==1.0.2
frozendict==2.4.6
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.1
google-api-python-client==2.140.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
googlemaps==4.10.0
groq==0.11.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.24.5
idna==3.7
importlib_metadata==8.2.0
interface-meta==1.3.0
ipykernel==6.29.4
ipython==8.24.0
ipywidgets==8.1.3
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_widgets==3.0.11
kagglehub==0.3.10
kaleido==0.2.1
kiwisolver==1.4.5
langchain==0.3.2
langchain-community==0.3.1
langchain-core==0.3.14
langchain-groq==0.2.0
langchain-openai==0.2.4
langchain-text-splitters==0.3.0
langdetect==1.0.9
langsmith==0.1.131
linearmodels==6.1
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.5.0
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.1.0
multitasking==0.0.11
mypy-extensions==1.0.0
narwhals==1.27.1
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
numpy==2.2.3
oauthlib==3.2.2
ofxparse==0.21
olefile==0.47
openai==1.63.2
openpyxl==3.1.5
orjson==3.10.7
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.3
pandas-datareader==0.10.0
parso==0.8.4
patsy==0.5.6
pdfminer.six==20231228
pdfplumber==0.11.4
peewee==3.17.7
pexpect==4.9.0
pikepdf==9.5.2
pillow==10.3.0
platformdirs==4.2.2
plotly==6.0.0
premailer==3.10.0
prettytable==3.11.0
prompt-toolkit==3.0.43
proto-plus==1.24.0
protobuf==5.27.2
psutil==5.9.8
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAutoGUI==0.9.54
pycountry==24.6.1
pycparser==2.22
pydantic==2.11.1
pydantic-settings==2.5.2
pydantic_core==2.33.0
pydeck==0.9.1
pydub==0.25.1
pydynpd==0.2.1
PyGetWindow==0.0.9
Pygments==2.18.0
pyhdfe==0.2.0
PyMsgBox==1.0.9
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pyobjc-core==10.2
pyobjc-framework-Cocoa==10.2
pyobjc-framework-Quartz==10.2
pyparsing==3.1.2
pypdf==5.0.1
PyPDF2==3.0.1
pypdfium2==4.30.0
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2025.2.18
python-magic==0.4.27
python-oxmsg==0.0.2
pytweening==1.2.0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
RapidFuzz==3.12.2
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retrying==1.3.4
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
rubicon-objc==0.4.9
safetensors==0.4.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
selenium==4.23.1
setuptools==70.0.0
setuptools-scm==8.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.35
stack-data==0.6.3
statsmodels==0.14.4
streamlit==1.42.1
sympy==1.13.3
tenacity==8.4.1
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tornado==6.4
tqdm==4.66.5
traitlets==5.14.3
transformers==4.43.3
trio==0.25.1
trio-websocket==0.11.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.1
unstructured==0.17.2
unstructured-client==0.32.0
uritemplate==4.1.1
urllib3==2.2.2
watchdog==4.0.1
wcwidth==0.2.13
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
wheel==0.43.0
widgetsnbextension==4.0.11
wooldridge==0.4.4
wrapt==1.16.0
wsproto==1.2.0
xlrd==2.0.1
xyzservices==2024.9.0
yagmail==0.15.293
yarl==1.13.1
yfinance==0.2.49
youtube-transcript-api==0.6.2
zipp==3.19.2
